Refactoring program logic to increase testability
